Move more methods to Base class.
